refactor: split parser into multiple files
